v0.1

+++ domain/src/main/java/com/borzg/domain/model/Movie.kt
@@ -8,20 +8,20 @@ import com.google.gson.annotations.SerializedName
 data class Movie(
         @PrimaryKey @SerializedName("id") override val id: Int,
         @SerializedName("backdrop_path") val backdrop_path: String?,
-        @SerializedName("budget") val budget: Int,
+        @SerializedName("budget") val budget: Long,
         @SerializedName("imdb_id") val imdbId: String?,
         @SerializedName("original_language") val originalLanguage: String,
         @SerializedName("original_title") val original_title: String,
         @SerializedName("overview") val overview: String?,
-        @SerializedName("popularity") val popularity: Double,
+        @SerializedName("popularity") val popularity: Float,
         @SerializedName("poster_path") val posterPath: String?,
         @SerializedName("release_date") val releaseDate: String,
-        @SerializedName("revenue") val revenue: Int,
+        @SerializedName("revenue") val revenue: Long,
         @SerializedName("runtime") val runtime: Int?,
         @SerializedName("title") val title: String,
-        @SerializedName("vote_average") val vote_average: Double,
+        @SerializedName("vote_average") val vote_average: Float,
         @SerializedName("vote_count") val vote_count: Int
-) : CinemaElement(){
+) : CinemaElement() {
         @Ignore @SerializedName("production_countries") var productionCountries: List<Country>? = null
         @ColumnInfo(typeAffinity = ColumnInfo.BLOB) var poster: ByteArray? = null
         @ColumnInfo(typeAffinity = ColumnInfo.BLOB) var backdrop: ByteArray? = null
